requirements.txt
# 核心依赖
openai>=1.0.0
anthropic>=0.5.0
yt-dlp>=2023.0
openai-whisper
pydub>=0.25.1
psutil>=5.9.0
jieba>=0.42.1
aiohttp>=3.8.0
playwright>=1.40.0

# Web应用
Flask>=2.0.0

# 数据库
# sqlite3 is part of Python standard library

# 其他工具
requests>=2.28.0